Improve objseq.ObjSeq_ExecuteActionCommand 99.13% -> 99.33%
restart is a pure 0/1 boolean only ever compared == 0. Retail compares it
bare (cmpwi r4,0); the s8 spelling forced a spurious extsb r0,r4 before the
compare. Typing it int drops the extension and matches retail.

+++ src/main/objseq.c
@@ -1084,7 +1084,7 @@ int ObjSeq_ExecuteActionCommand(u8* obj, u8* action, u8** cmdPtr, s8 flags, void
     u8* entry;
     int opcode;
     int sub;
-    s8 restart;
+    int restart;
     int reps;
     int val;
     int slot;
